backend/requirements.txt

# --- Doğrulama & Serileştirme ---
# --- Hız Sınırlama ---
# --- Kimlik Doğrulama ---
# --- Loglama ---
# --- Redis ---
# --- Test ---
# --- Veritabanı ---
# --- Web Framework ---
# --- WebSocket ---
# --- Yardımcı ---
# ============================================
# TOPLANTI UYGULAMASI - PYTHON BAĞIMLILIKLARI
alembic==1.14.0
asyncpg==0.30.0
bcrypt==4.2.1
email-validator==2.2.0
fastapi==0.115.6
greenlet==3.1.1
httpx==0.28.1
loguru==0.7.3
orjson==3.10.12
passlib[bcrypt]==1.7.4
pydantic-settings==2.7.0
pydantic==2.10.3
pytest-asyncio==0.25.0
pytest==8.3.4
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.18
redis[hiredis]==5.2.1
slowapi==0.1.9
sqlalchemy[asyncio]==2.0.36
uvicorn[standard]==0.34.0
websockets==14.1
